Reset to ea3d26febc2c995a75b5e2b837d434ebeefd8d72 (PR #1463)

+++ app/service_functions/rdv_updater.rb
@@ -2,43 +2,47 @@
 
 module RdvUpdater
   class << self
-    def update(author, rdv, rdv_params)
+    def update_by_user(rdv, rdv_params)
+      update(rdv, rdv_params, :user)
+    end
+
+    def update_by_agent(rdv, rdv_params)
+      update(rdv, rdv_params, :agent)
+    end
+
+    private
+
+    def update(rdv, rdv_params, by)
       rdv.updated_at = Time.zone.now
       # TODO: replace this manual touch. It forces creating a version when an
       # agent or a user is removed from the RDV. the touch: true option on the
       # association does not do it for some reason I could not figure out
 
       if rdv_params[:status].present?
-        rdv_params[:cancelled_at] = rdv_params[:status].in?(%w[excused notexcused]) ? Time.zone.now : nil
+        rdv_params[:cancelled_at] = cancel_status?(rdv_params[:status]) ? Time.zone.now : nil
       end
 
-      result = rdv.update(rdv_params)
-      send_relevant_notifications(author, rdv, rdv_params) if result
-      result
+      return false unless rdv.update(rdv_params)
+
+      rdv.file_attentes.destroy_all if cancel_status?(rdv_params[:status])
+      notify_cancelled(rdv, by) if rdv_params[:status] == "excused"
+      true
     end
 
-    def send_relevant_notifications(author, rdv, rdv_params)
-      notify_cancellation_to_agent(author, rdv) if rdv_params[:status].in? %w[excused notexcused]
-      notify_cancellation_to_user(author, rdv) if rdv_params[:status] == "excused"
-      notify_starts_at_change(author, rdv) if rdv_params[:starts_at].present?
+    def notify_cancelled(rdv, by)
+      send("notify_#{by}", rdv)
     end
 
-    def notify_cancellation_to_agent(author, rdv)
-      rdv.file_attentes.destroy_all
-      Notifications::Rdv::RdvCancelledService.perform_with(rdv, author)
+    def notify_agent(rdv)
+      Notifications::Rdv::RdvCancelledByAgent.perform_with(rdv)
     end
 
-    def notify_cancellation_to_user(author, rdv)
-      case author
-      when Agent
-        Notifications::Rdv::RdvCancelledByAgent.perform_with(rdv, author)
-      when User
-        Notifications::Rdv::RdvCancelledByUser.perform_with(rdv, author)
-      end
+    def notify_user(rdv)
+      Notifications::Rdv::RdvCancelledByUser.perform_with(rdv)
     end
 
-    def notify_starts_at_change(author, rdv)
-      Notifications::Rdv::RdvDateUpdatedService.perform_with(rdv, author)
+    def cancel_status?(status)
+      %w[excused notexcused].include?(status)
     end
   end
 end
